Move responsive video player stuff into it's own component and use it on the projects list page as well.
