+++ web/src/components/project.tsx
@@ -1,27 +1,19 @@
-import React, { useEffect, useRef, useState } from 'react';
-import ReactPlayer from 'react-player';
-import Container from './container';
+import React from 'react';
+import styled, { css } from 'styled-components';
 import { buildImageObj } from '../lib/helpers';
 import { imageUrlFor } from '../lib/image-url';
-import SimpleReactLightbox, { SRLWrapper } from 'simple-react-lightbox';
+import { Link } from 'gatsby';
+import SimpleReactLightbox from 'simple-react-lightbox';
+import { DeviceWidth } from '../styles/mediaQueries';
 
-import FullscreenIcon from '../images/fullscreen-icon.svg';
-import ArrowRight from '../images/arrow-right.svg';
-
-import styled, { css } from 'styled-components';
 import { font } from '../styles/typography';
 import { theme } from '../styles/theme';
-import PhotoGrid from './photo-grid';
-import { DeviceWidth } from '../styles/mediaQueries';
-import useWindowDimensions from '../hooks/useWindowDimensions';
-import { Link } from 'gatsby';
 
-function preventScroll(e) {
-  console.log('preventing scroll!');
-  e.preventDefault();
-  e.stopPropagation();
-  return false;
-}
+import Container from './container';
+import PhotoGrid from './photo-grid';
+import ResponsiveVideoContainer, { VideoContainer } from './ResponsiveVideoContainer';
+
+import ArrowRight from '../images/arrow-right.svg';
 
 export interface ProjectProps {
   project: any;
@@ -33,13 +25,6 @@ function Project(props: ProjectProps) {
   const { project, linkToPrevious, linkToNext } = props;
   const { clientName, clientLogo, projectDate, title, videoUrl, excerpt, contributors } = project;
   const images: Array<any> = project.image;
-  const { width: screenWidth } = useWindowDimensions();
-  const [videoWidth, setVideoWidth] = useState('100%');
-
-  useEffect(() => {
-    const w = screenWidth > 1120 ? '100%' : '150%';
-    setVideoWidth(w);
-  }, [screenWidth]);
 
   function generateContributorString(contributor) {
     let contributorString = '';
@@ -65,67 +50,11 @@ function Project(props: ProjectProps) {
     </TitleContainer>
   );
 
-  const [playing, setPlaying] = useState(true);
-  const [isFullscreen, setIsFullscreen] = useState(false);
-
-  const player = useRef(null);
-
-  const handleClickFullscreen = () => {
-    setPlaying(false);
-    setIsFullscreen(true);
-    document.querySelector('#scrollable')?.addEventListener('wheel', preventScroll);
-    document.querySelector('#modalVideoContainer')?.addEventListener('wheel', preventScroll);
-  };
-
-  const handleExitFullscreen = () => {
-    setPlaying(true);
-    setIsFullscreen(false);
-    document.querySelector('#scrollable')?.removeEventListener('wheel', preventScroll);
-  };
-
   return (
     <article id="scrollable">
       <SimpleReactLightbox>
         {videoUrl && (
-          <>
-            {isFullscreen && (
-              <ModalDiv onClick={handleExitFullscreen} onScroll={preventScroll}>
-                <ModalVideoContainer id="modalVideoContainer">
-                  <ReactPlayer
-                    style={{ position: 'absolute', top: 0, left: 0 }}
-                    url={videoUrl}
-                    controls={true}
-                    config={{
-                      vimeo: {
-                        playerOptions: { responsive: true },
-                      },
-                    }}
-                    width="100%"
-                    height="100%"
-                  />
-                </ModalVideoContainer>
-              </ModalDiv>
-            )}
-            <VideoContainer>
-              <ReactPlayer
-                ref={player}
-                style={{ position: 'absolute', top: 0, left: 0 }}
-                url={videoUrl}
-                controls={false}
-                playing={playing}
-                config={{
-                  vimeo: {
-                    playerOptions: { background: true, loop: true, responsive: true },
-                  },
-                }}
-                playsinline
-                width={videoWidth}
-                height="100%"
-              />
-              {titleBlock}
-              <FullscreenButton onClick={handleClickFullscreen} aria-label="Maximize video" />
-            </VideoContainer>
-          </>
+          <ResponsiveVideoContainer videoUrl={videoUrl}>{titleBlock}</ResponsiveVideoContainer>
         )}
 
         {!videoUrl && images.length && (
@@ -181,71 +110,6 @@ function Project(props: ProjectProps) {
 }
 
 export default Project;
-
-const ModalDiv = styled.div`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.9);
-  z-index: 9999;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-const ModalVideoContainer = styled.div`
-  position: relative;
-  width: 70%;
-  height: 0;
-  padding-top: 40%;
-  overflow: hidden;
-`;
-
-const VideoContainer = styled.div`
-  position: relative;
-  border-bottom-left-radius: 35px;
-  border-bottom-right-radius: 35px;
-  overflow: hidden;
-  box-shadow: 0px 7px 16px 0px #0000004f;
-
-  @media (${DeviceWidth.mediaMinSmall}) {
-    height: 600px;
-    /* padding-top: 56.25%; */
-  }
-
-  @media (max-width: 1120px) {
-    & > div > div > div {
-      margin-left: -25%;
-    }
-  }
-
-  @media (min-width: 450px) and (max-width: 750px) {
-    height: unset;
-    padding-top: 80%;
-  }
-
-  @media (${DeviceWidth.mediaMaxSmall}) {
-    height: 34vh;
-    border-bottom-left-radius: 20px;
-    border-bottom-right-radius: 20px;
-  }
-`;
-
-const FullscreenButton = styled.button`
-  position: absolute;
-  bottom: 50px;
-  right: 50px;
-  z-index: 998;
-  background: none;
-  background-image: url(${FullscreenIcon});
-  background-repeat: no-repeat;
-  background-size: contain;
-  border: none;
-  height: 30px;
-  width: 30px;
-`;
 
 const MainPhotoContainer = styled(VideoContainer)`
   display: flex;
