+++ web/src/pages/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { graphql, Link } from 'gatsby';
 import { mapEdgesToNodes, filterOutDocsWithoutSlugs } from '../lib/helpers';
 import GraphQLErrorList from '../components/graphql-error-list';
@@ -11,9 +11,8 @@ import Layout from '../containers/layout';
 
 import ArrowRight from '../images/arrow-right.svg';
 import { theme } from '../styles/theme';
-import ReactPlayer from 'react-player';
 import { DeviceWidth } from '../styles/mediaQueries';
-import useWindowDimensions from '../hooks/useWindowDimensions';
+import ResponsiveVideoContainer from '../components/ResponsiveVideoContainer';
 
 export const query = graphql`
   query ProjectsPageQuery {
@@ -56,34 +55,11 @@ const ProjectsPage = (props) => {
 
   const videoUrl = data?.site?.videoUrl;
 
-  const { width: screenWidth } = useWindowDimensions();
-  const [videoWidth, setVideoWidth] = useState('100%');
-
-  useEffect(() => {
-    const w = screenWidth > 675 ? '100%' : '240%';
-    setVideoWidth(w);
-  }, [screenWidth]);
-
   return (
     <Layout>
       <SEO title="Work" />
 
-      <VideoContainer>
-        <ReactPlayer
-          style={{ position: 'absolute', top: 0, left: 0 }}
-          url={videoUrl}
-          // playing={true}
-          // controls={true}
-          config={{
-            vimeo: {
-              playerOptions: { background: true, loop: true, responsive: true },
-            },
-          }}
-          playsinline
-          width={videoWidth}
-          height="100%"
-        />
-      </VideoContainer>
+      <ResponsiveVideoContainer fullHeight videoUrl={videoUrl} />
 
       <Container wide short grow>
         {projectNodes && projectNodes.length > 0 && (
@@ -108,28 +84,6 @@ const ProjectsPage = (props) => {
 };
 
 export default ProjectsPage;
-
-const VideoContainer = styled.div`
-  position: relative;
-  border-bottom-left-radius: 35px;
-  border-bottom-right-radius: 35px;
-  overflow: hidden;
-  box-shadow: 0px 7px 16px 0px #0000004f;
-
-  @media (${DeviceWidth.mediaMinMedium}) {
-    padding-top: 56.25%;
-  }
-
-  @media (${DeviceWidth.mediaMaxMedium}) {
-    height: 60vh;
-    border-bottom-left-radius: 20px;
-    border-bottom-right-radius: 20px;
-
-    & > div > div > div {
-      margin-left: -59%;
-    }
-  }
-`;
 
 const Title = styled.h1`
   ${font('title24')};
